Add support for internal pull-up resistor
On platforms such as Arduino, a new #define can now be set in OneWire.h
to enable the internal pull-up resistor. This means that no external
resistor is needed to search for 1-wire devices. Do not expect this to
work for temperature sensors and other such devices, that have higher
power requirements than simple serial number devices.