Update OneWire_direct_gpio.h
Use the VPORTx registers - should give slightly smaller, faster, atomic code. Previous implementation used the PORTx registers; those are not in the I/O space, so |= and &= were not atomic! They still may not be in all cases (depending on what the compiler is able to optimize away) - but this will transparently take advantage of updates to the macros supplied with the cores which do, if/when they come.