Update OneWire.cpp to correct the ESP32 code
The ESP32 “define noInterrupts()…” code does not work (they do NOT replace the “noInterrupts();” commands), and also contain unbalanced { brackets anyway.  It is also unnecessarily duplicated in the .h and .cpp files.  In addition this change is NOT required for any of this single-core ESP32 units.  

This PR fixes the above, including adding the required
 static portMUX_TYPE OneWire_mutex = portMUX_INITIALIZER_UNLOCKED;